Board State now showing correctly It first reads in from a file which is opened by a ajax function with javascript Then it parses the file into the data format we want
Right now the turn is set to 0 and not changed.

Thus it is redering the first turn's board.

If you modify the turn number it will change accordingly to reflect what turn data it has
The GameState object is a list of all turns
Each turn in GameState has an object that has information about turn number, player's name which is p1 and p2, also board state which is Board, at last we have comment which is comment.

Javascript opening a file.... is a nightmare.